Remove override of proposal filters that has changed in decidim in the meantime
The original intention of this override was to always show the "origin"
filter, even when official proposals are disabled.
Original issue: https://github.com/puzzle/quartieridee/issues/25
(cell I108 in the linked google spreadsheet)

But that template has changed a lot since mid 2020 and it is now relied
upon for showing a spinner while the filters are evaluating. If the
template is not the way this spinner script expects, javascript will
crash while filtering, and the proposal filters don't work.

Fixes #196