Use JS to fix over-eager erb escaping
For some reason, the ERB compiler will url encode anything inside of a
href attribute of an html link in this application.
I was not able to debug the erb compiler, and a previous fix overriding
the template broke the proposal filters. So I resort to using javascript
for correcting the placeholder href="${link}".

Fixes #108