Fix proposal creation map with drag marker
Fixes #269

Decidim added a new type of map controller for the draggable marker map.
So our custom maps now have to extend either the classic
MapMarkersController or the new DragMarkerController. For this we need a
new composable pattern for our customizations.

Also, our photon geocoder fix does not seem to take effect anymore, so
we resort to a different fix for the swapped latitude and longitude.